Refactoring integrations tests and functionality

+++ transfer-service/src/main/java/com/yellowpepper/transferservice/dtos/Transfer.java
@@ -39,7 +39,7 @@ public class Transfer {
 
     private Float taxCollected;
 
-    private Double cad;
+    private float cad;
